update snap point calculation system and camera movement system
By only storing snap point information through pure data in the buffer instead of storing it through entities as before, it avoids having to retrieve non-continuous query data and minimizes the number of unnecessary entities in the project (retrieving information from different entities), which reduces performance and is difficult to maintain and debug. Update the data system to match this data processing direction because this requires synchronization and data conversion between ghost and build during the spawn new build phase.
Update the camera movement system to match the operation during construction
Update libraries and unity versions